OSIS-4817 : Ajout de le liste des formations avec les données administratives
