+++ program_management/templates/search.html
@@ -130,6 +130,14 @@
                                         &nbsp;{% trans 'List of trainings' %}
                                     </a>
                                 </li>
+                                <li>
+                                    <a id="btn_produce_xls_administrative_data"
+                                       class="download"
+                                       style="margin-right:10px;"
+                                       title="{% trans 'Produce xls of trainings with administrative data' %}">
+                                        &nbsp;{% trans 'List of trainings with administrative data' %}
+                                    </a>
+                                </li>
                             </ul>
                             <form id="download_xls" action="{{ request.path }}" method="post">
                                 {% csrf_token %}
@@ -184,17 +192,18 @@
             {{items_per_page}}, "{% url 'version_program' %}", columnDefs);
         });
         // FIXME remove xls_order_col / xls_order ==> USE ordering params
-        $("#btn_produce_xls").click(function (e) {
+        $("#btn_produce_xls, #btn_produce_xls_administrative_data").click(function (e) {
 
             let table = $('#table_education_groups').DataTable();
             let order = table.order();
             $("#xls_order_col").val(get_ordering_column(order));
-            alert(order[0][1]);
             $("#xls_order").val(order[0][1]);
 
             let action_value;
             if ($(this).attr('id') === 'btn_produce_xls') {
                 action_value = 'xls';
+            } else if ( $(this).attr('id') === 'btn_produce_xls_administrative_data' ) {
+                action_value = 'xls_administrative';
             }
             prepare_xls(e, action_value);
             $("#xls_order").val('');
