+++ program_management/views/search.py
@@ -34,7 +34,7 @@ from django.utils.html import format_html
 from django.utils.translation import gettext_lazy as _
 from django_filters.views import FilterView
 
-from base.business.education_group import create_xls, ORDER_COL, ORDER_DIRECTION
+from base.business.education_group import create_xls, ORDER_COL, ORDER_DIRECTION, create_xls_administrative_data
 from base.forms.search.search_form import get_research_criteria
 from base.models.academic_year import starting_academic_year
 from base.models.education_group_type import EducationGroupType
@@ -61,7 +61,18 @@ def _create_xls(view_obj, context, **response_kwargs):
     return create_xls(user, egys, filters, order)
 
 
+def _create_xls_administrative_data(view_obj, context, **response_kwargs):
+    user = view_obj.request.user
+    egys = context["filter"].qs
+    filters = _get_filter(context["form"])
+    # FIXME: use ordering args in filter_form! Remove xls_order_col/xls_order property
+    order = {ORDER_COL: view_obj.request.GET.get('xls_order_col'),
+             ORDER_DIRECTION: view_obj.request.GET.get('xls_order')}
+    return create_xls_administrative_data(user, egys, filters, order)
+
+
 @RenderToExcel("xls", _create_xls)
+@RenderToExcel("xls_administrative", _create_xls_administrative_data)
 class EducationGroupSearch(LoginRequiredMixin, PermissionRequiredMixin, CacheFilterMixin, SearchMixin, FilterView):
     model = GroupYear
     template_name = "search.html"
